fix: don't use bigints (sometimes it's out of bounds) (#3781)

+++ warehouse/oso_sqlmesh/models/staging/superchain/stg_superchain__4337_userop_logs.sql
@@ -35,8 +35,8 @@ SELECT
   sender AS sender_address,
   paymaster AS paymaster_address,
   contract_address,
-  actualgascost::BIGINT AS userop_gas_price,
-  actualgasused::BIGINT AS userop_gas_used,
+  actualgascost::DOUBLE AS userop_gas_price,
+  actualgasused::DOUBLE AS userop_gas_used,
   @chain_name(chain) AS chain
 FROM @oso_source(
   'bigquery.optimism_superchain_4337_account_abstraction_data.useroperationevent_logs_v2'
